// types/MockChainlinkAggregator.d.ts
/* Generated by ts-generator ver. 0.0.8 */
/* tslint:disable */

import { Contract, ContractTransaction, EventFilter, Signer } from "ethers";
import { Listener, Provider } from "ethers/providers";
import { Arrayish, BigNumber, BigNumberish, Interface } from "ethers/utils";
import {
  TransactionOverrides,
  TypedEventDescription,
  TypedFunctionDescription
} from ".";

interface MockChainlinkAggregatorInterface extends Interface {
  functions: {
    latestAnswer: TypedFunctionDescription<{ encode([]: []): string }>;
  };

  events: {};
}

export class MockChainlinkAggregator extends Contract {
  connect(
    signerOrProvider: Signer | Provider | string
  ): MockChainlinkAggregator;
  attach(addressOrName: string): MockChainlinkAggregator;
  deployed(): Promise<MockChainlinkAggregator>;

  on(event: EventFilter | string, listener: Listener): MockChainlinkAggregator;
  once(
    event: EventFilter | string,
    listener: Listener
  ): MockChainlinkAggregator;
  addListener(
    eventName: EventFilter | string,
    listener: Listener
  ): MockChainlinkAggregator;
  removeAllListeners(eventName: EventFilter | string): MockChainlinkAggregator;
  removeListener(eventName: any, listener: Listener): MockChainlinkAggregator;

  interface: MockChainlinkAggregatorInterface;

  functions: {
    latestAnswer(): Promise<BigNumber>;
  };

  latestAnswer(): Promise<BigNumber>;

  filters: {};

  estimate: {
    latestAnswer(): Promise<BigNumber>;
  };
}